requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
waitress==2.1.2
opencv-python==4.8.1.78
face-recognition==1.3.0
mediapipe==0.10.8
numpy==1.26.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
pillow==10.1.0
psycopg2-binary==2.9.9